refactorof of target deleting

+++ targets/src/main/java/be/vinci/ipl/projet2024/group07/targets/TargetsService.java
@@ -2,10 +2,12 @@ package be.vinci.ipl.projet2024.group07.targets;
 
 import be.vinci.ipl.projet2024.group07.targets.models.Server;
 import be.vinci.ipl.projet2024.group07.targets.models.Target;
+import be.vinci.ipl.projet2024.group07.targets.repositories.AttacksProxy;
 import be.vinci.ipl.projet2024.group07.targets.repositories.ServersProxy;
 import be.vinci.ipl.projet2024.group07.targets.repositories.TargetsRepository;
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -13,12 +15,14 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class TargetsService {
-    private TargetsRepository targetsRepository;
+    private final TargetsRepository targetsRepository;
     private final ServersProxy serversProxy;
+    private final AttacksProxy attacksProxy;
 
-    public TargetsService(TargetsRepository targetsRepository, ServersProxy serversProxy) {
+    public TargetsService(TargetsRepository targetsRepository, ServersProxy serversProxy, AttacksProxy attacksProxy) {
         this.targetsRepository = targetsRepository;
         this.serversProxy = serversProxy;
+        this.attacksProxy = attacksProxy;
     }
 
     public Target createTarget(Target target) {
@@ -43,8 +47,12 @@ public class TargetsService {
         return targetsRepository.findAllByServersBetween(min, max);
     }
 
-    public void deleteTarget(int id) {
+    public boolean deleteTarget(int id) {
+        if (!targetsRepository.existsById(id)) return false;
         targetsRepository.deleteById(id);
+        serversProxy.deleteByTarget(id);
+        attacksProxy.deleteTargets(id);
+        return true;
     }
 
     public boolean updateOne(Target target){
@@ -54,18 +62,18 @@ public class TargetsService {
     }
 
     public Iterable<String> getIpColocated() {
-        Map<String, Set<Target>> dicoIp = new HashMap();
+        System.out.println("test");
+        Map<String, Set<Target>> dicoIp = new HashMap<>();
         for (Target target : targetsRepository.findAll()) {
             Iterable<Server> servers = serversProxy.readByTarget(target.getId());
+            System.out.println(servers);
             for (Server server : servers) {
-                if (dicoIp.containsKey(server.getIpAddress())) {
-                    dicoIp.get(server.getIpAddress()).add(target);
-                } else {
-                    Set<Target> targets = Set.of(target);
-                    dicoIp.put(server.getIpAddress(), targets);
-                }
+                Set<Target> targetSet = dicoIp.computeIfAbsent(server.getIpAddress(), k -> new HashSet<>());
+                targetSet.add(target);
             }
         }
+
+        System.out.printf("dicoIp: %s\n", dicoIp);
 
         ArrayList<String> allIp = new ArrayList<>();
         for (String ip : dicoIp.keySet()) {
